+++ Changelog.md
@@ -6,6 +6,7 @@
 - Fix quest flag expressions are no longer case sensitive
 - Fix autovendor race condition with small autovendor profiles
 - Add warning when using ub tools that require loot profiles, and vit has a profile loaded.
+- Add ability to tag characters, and broadcast commands to specific tags with `/ub bct` command
 
 ## 0.2.3 (2021-01-30) [UtilityBeltInstaller-0.2.3.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/fee6e8ec9b64ef0aa33535e98fe3ed3d/UtilityBeltInstaller-0.2.3.exe)
 - Fix `cleargvar[]` expression wasn't actually clearing vars